move ps to feynml
